refactor: componentize Trilha for reusable subject trails
- Extract TrailPath component (trail-path.tsx): reusable trail with
  configurable units/nodes, exports TrailUnit/TrailNode types
- Extract SubjectHeader component (subject-header.tsx): shows selected
  subject icon/name, "Todos os assuntos" dropdown (→ /materias),
  filter button (→ /filtro-assuntos)
- Create Filtro Assuntos modal (filtro-assuntos.tsx): transparentModal
  bottom sheet with numbered subject list (Citologia selected/green,
  Genética, Ecologia, Evolução)
- Refactor trilha.tsx to use TrailPath + SubjectHeader with Biologia
  data, making it easy to add more subjects
- Register filtro-assuntos in root layout as transparentModal
